Support for channel state control
Allows to set the channel state. The Mirth API comes w/ some (partially strange) limitations. It e.g. does not allow to directly bring a channel to the PAUSED state when it is stopped. In contrary to that it is possible to deploy a channel directly to PAUSED state.

This implementation bypasses these limitations and allows to bring mirth from any state to the desired state. This, however, comes at a price: If a channel should be set from STOPPED to PAUSED or should be deployed to an arbitrary state that is not configured as the initial state of the channel, the Mirth Administrator will indicate a revision change. This is due to the way the limitations are circumvented. (initial state is changed to the desired state, channel is deployed, initial state is changed back to the original state which causes the revision change)